Fully implement this, adapt test assertions
