+++ aleph/tests/test_entities_api.py
@@ -643,22 +643,17 @@ class EntitiesApiTestCase(TestCase):
                                       headers=headers,
                                       content_type='application/json')
 
-        edge_types = [registry.name.name, registry.email.name,
-                      registry.identifier.name, registry.iban.name,
-                      registry.phone.name, registry.address.name,
-                      registry.url.name, registry.checksum.name,
-                      registry.entity.name]
-        query_string = '&'.join('edge_types=' + t for t in edge_types)
+        query_string = '&'.join('edge_types=' + t.name for t in registry.pivots)  # noqa
 
         url = '/api/2/entities/%s/expand?%s&limit=0' % (person1.json['id'], query_string)  # noqa
         stats = self.client.get(url, headers=headers)
         assert stats.status_code == 200, (stats.status_code, stats.json)
         validate(stats.json, 'QueryResponse')
-        assert stats.json['total'] == 4, stats.json
+        assert stats.json['total'] == 5, stats.json
         results = stats.json['results']
         for result in results:
             validate(result, 'EntityExpand')
-            assert result['property'] in ('holder', 'Ownership', 'email'), stats.json  # noqa
+            assert result['property'] in ('passport', 'name', 'ownershipOwner', 'email'), stats.json  # noqa
             if result['property'] == 'email':
                 assert result['count'] == 2, results
             else:
@@ -668,22 +663,24 @@ class EntitiesApiTestCase(TestCase):
         res = self.client.get(url, headers=headers)
         assert res.status_code == 200, (res.status_code, res.json)
         validate(res.json, 'EntityExpand')
-        assert res.json['total'] == 4, pformat(res.json)
+        assert res.json['total'] == 5, pformat(res.json)
         results = res.json['results']
-        assert len(results) == 3, pformat(results)
+        assert len(results) == 4, pformat(results)
         for res in results:
             prop = res['property']
-            assert prop in ('email', 'holder', 'Ownership'), results
+            assert prop in ('passport', 'name', 'ownershipOwner', 'email'), results  # noqa
             if prop == 'email':
                 assert res['count'] == 2
                 for ent in res['entities']:
                     assert ent['name'] in ('Undercover Osama', 'John Doe')
-            if prop == 'Ownership':
+            if prop == 'ownershipOwner':
                 assert res['count'] == 1
-                assert res['entities'][0]['name'] == 'Al-Qaeda'
-            if prop == 'holder':
+                assert len(res['entities']) == 2
+                for nested in res['entities']:
+                    assert nested['schema'] in ('Ownership', 'Company'), nested
+            if prop == 'passport':
                 assert res['count'] == 1
-                assert res['entities'][0]['name'] == 'A1B2C3'
+                assert res['entities'][0]['schema'] == 'Passport', res
         url = '/api/2/entities/%s/expand?%s' % (company1.json['id'], query_string)  # noqa
         res = self.client.get(url, headers=headers)
         assert res.status_code == 200, (res.status_code, res.json)
@@ -693,6 +690,6 @@ class EntitiesApiTestCase(TestCase):
         assert len(results) == 1, pformat(results)
         for res in results:
             prop = res['property']
-            assert prop == 'Ownership'
+            assert prop in ('Ownership', 'name'), prop
             assert res['count'] == 1
-            assert res['entities'][0]['name'] == 'Osama bin Laden'
+            # assert res['entities'][0]['name'] == 'Osama bin Laden'
