+++ services/ingest-file/requirements.txt
@@ -3,7 +3,7 @@ normality==2.0.0
 pantomime==0.4.0
 servicelayer[google,amazon]==1.10.2
 balkhash[leveldb,sql]==1.2.1
-followthemoney==1.29.3
+followthemoney==1.29.4
 languagecodes==1.0.5
 countrytagger==0.1.2
 psycopg2-binary==2.8.5
